[FEAT] Add controllers

+++ src/schemas/media_creators.ts
@@ -1,6 +1,6 @@
 import * as z from 'zod';
 
 export const MediaCreatorSchema = z.object({
-    media_id: z.number(),
-    creator_id: z.number(),
+    media_id: z.number().int(),
+    creator_id: z.number().int(),
 }).strip();
